app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0

google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
isodate==0.6.1

boto3==1.34.162
botocore==1.34.162